DeferredFolds.UnfoldlM

Documentation

newtype UnfoldlM m a

A monadic variation of DeferredFolds.Unfoldl

Constructors

UnfoldlM (forall x. (x -> a -> m x) -> x -> m x)

filter :: Monad m => (a -> m Bool) -> UnfoldlM m a -> UnfoldlM m a

Filter the values given a predicate

unfoldr :: Monad m => Unfoldr a -> UnfoldlM m a

null :: Monad m => UnfoldlM m input -> m Bool

Check whether it's empty

fold :: Fold input output -> UnfoldlM Identity input -> output

Apply a Gonzalez fold

mapM_ :: Monad m => (input -> m ()) -> UnfoldlM m input -> m ()

A more efficient implementation of mapM_

forM_ :: Monad m => UnfoldlM m input -> (input -> m ()) -> m ()

Same as mapM_ with arguments flipped

foldM :: Monad m => FoldM m input output -> UnfoldlM m input -> m output

Apply a monadic Gonzalez fold

hoist :: (forall a. m a -> n a) -> (forall a. n a -> m a) -> UnfoldlM m a -> UnfoldlM n a

Change the base monad using invariant natural transformations

foldlM' :: Monad m => (output -> input -> m output) -> output -> UnfoldlM m input -> m output

Perform a monadic strict left fold

foldable :: (Monad m, Foldable foldable) => foldable a -> UnfoldlM m a

Construct from any foldable

mapFoldMInput :: Monad m => (forall x. FoldM m b x -> FoldM m a x) -> UnfoldlM m a -> UnfoldlM m b

Lift a fold input mapping function into a mapping of unfolds

foldlRunner :: Monad m => (forall x. (x -> a -> x) -> x -> x) -> UnfoldlM m a

Construct from a specification of how to execute a left-fold

foldrRunner :: Monad m => (forall x. (a -> x -> x) -> x -> x) -> UnfoldlM m a

Construct from a specification of how to execute a right-fold

intsInRange :: Monad m => Int -> Int -> UnfoldlM m Int

Ints in the specified inclusive range

tVarValue :: TVar a -> UnfoldlM STM a

TVar contents

byteStringBytes :: ByteString -> UnfoldlM IO Word8

Bytes of a bytestring

shortByteStringBytes :: Monad m => ShortByteString -> UnfoldlM m Word8

Bytes of a short bytestring

primArray :: (Monad m, Prim prim) => PrimArray prim -> UnfoldlM m prim

Elements of a prim array

primArrayWithIndices :: (Monad m, Prim prim) => PrimArray prim -> UnfoldlM m (Int, prim)

Elements of a prim array coming paired with indices