DeferredFolds.Unfoldr

Documentation

newtype Unfoldr a

A projection on data, which only knows how to execute a right-fold.

It is a monad and a monoid, and is very useful for efficiently aggregating the projections on data intended for right-folding, since its concatenation (<>) has complexity of O(1).

Intuition

The intuition of what this abstraction is all about can be derived from lists.

Let's consider the foldr function for lists:

foldr :: (a -> b -> b) -> b -> [a] -> b

If we rearrange its parameters we get

foldr :: [a] -> (a -> b -> b) -> b -> b

Which in Haskell is essentially the same as

foldr :: [a] -> (forall b. (a -> b -> b) -> b -> b)

We can isolate that part into an abstraction:

newtype Unfoldr a = Unfoldr (forall b. (a -> b -> b) -> b -> b)

Then we get to this simple morphism:

list :: [a] -> Unfoldr a
list list = Unfoldr (\ step init -> foldr step init list)

We can do the same with say Data.Text.Text:

text :: Text -> Unfoldr Char
text text = Unfoldr (\ step init -> Data.Text.foldr step init text)

And then we can use those both to concatenate with just an O(1) cost:

abcdef :: Unfoldr Char
abcdef = list ['a', 'b', 'c'] <> text "def"

Please notice that up until this moment no actual data materialization has happened and hence no traversals have appeared. All that we've done is just composed a function, which only specifies which parts of data structures to traverse to perform a right-fold. Only at the moment where the actual folding will happen will we actually traverse the source data. E.g., using the "fold" function:

abcdefLength :: Int
abcdefLength = fold Control.Foldl.length abcdef

Constructors

Unfoldr (forall x. (a -> x -> x) -> x -> x)

vector :: Vector vector a => vector a -> Unfoldr a

Elements of a vector

zipWith :: (a -> b -> c) -> Unfoldr a -> Unfoldr b -> Unfoldr c

filter :: (a -> Bool) -> Unfoldr a -> Unfoldr a

Filter the values given a predicate

takeWhile :: (a -> Bool) -> Unfoldr a -> Unfoldr a

take :: Int -> Unfoldr a -> Unfoldr a

reverse :: Unfoldr a -> Unfoldr a

Reverse the order.

Use with care, because it requires to allocate all elements.

fold :: Fold input output -> Unfoldr input -> output

Apply a Gonzalez fold

intersperse :: a -> Unfoldr a -> Unfoldr a

Insert a separator value between each element.

Behaves the same way as intersperse.

foldM :: Monad m => FoldM m input output -> Unfoldr input -> m output

Apply a monadic Gonzalez fold

cons :: a -> Unfoldr a -> Unfoldr a

snoc :: a -> Unfoldr a -> Unfoldr a

foldable :: Foldable foldable => foldable a -> Unfoldr a

Construct from any foldable

intsInRange :: Int -> Int -> Unfoldr Int

Ints in the specified inclusive range

byteStringBytes :: ByteString -> Unfoldr Word8

Bytes of a bytestring

shortByteStringBytes :: ShortByteString -> Unfoldr Word8

Bytes of a short bytestring

primArray :: Prim prim => PrimArray prim -> Unfoldr prim

Elements of a prim array

primArrayWithIndices :: Prim prim => PrimArray prim -> Unfoldr (Int, prim)

Elements of a prim array coming paired with indices

mapAssocs :: Map key value -> Unfoldr (key, value)

Associations of a map

intMapAssocs :: IntMap value -> Unfoldr (Int, value)

Associations of an intmap

intSet :: IntSet -> Unfoldr Int

Elements of IntSet.

enumsFrom :: Enum a => a -> Unfoldr a

Ascending infinite stream of enums starting from the one specified

enumsInRange :: (Enum a, Ord a) => a -> a -> Unfoldr a

Enums in the specified inclusive range

intsFrom :: Int -> Unfoldr Int

Ascending infinite stream of ints starting from the one specified

hashMapKeys :: HashMap key value -> Unfoldr key

Keys of a hash-map

hashMapAssocs :: HashMap key value -> Unfoldr (key, value)

Associations of a hash-map

hashMapAt :: (Hashable key, Eq key) => HashMap key value -> key -> Unfoldr value

Value of a hash-map by key

hashMapValue :: (Hashable key, Eq key) => key -> HashMap key value -> Unfoldr value

Deprecated: Use hashMapAt instead

Value of a hash-map by key

hashMapValues :: (Hashable key, Eq key) => HashMap key value -> Unfoldr key -> Unfoldr value

Values of a hash-map by their keys

vectorWithIndices :: Vector vector a => vector a -> Unfoldr (Int, a)

Elements of a vector coming paired with indices

binaryDigits :: Integral a => a -> Unfoldr a

Binary digits of a non-negative integral number.

reverseBinaryDigits :: Integral a => a -> Unfoldr a

Binary digits of a non-negative integral number in reverse order.

reverseDigits

Arguments

:: Integral a

=> a
Radix

-> a
Number

-> Unfoldr a

Digits of a non-negative number in numeral system based on the specified radix. The digits come in reverse order.

E.g., here's how an unfold of binary digits in proper order looks:

binaryDigits :: Integral a => a -> Unfoldr a
binaryDigits = reverse . reverseDigits 2

octalDigits :: Integral a => a -> Unfoldr a

Octal digits of a non-negative integral number.

reverseOctalDigits :: Integral a => a -> Unfoldr a

Octal digits of a non-negative integral number in reverse order.

decimalDigits :: Integral a => a -> Unfoldr a

Decimal digits of a non-negative integral number.

reverseDecimalDigits :: Integral a => a -> Unfoldr a

Decimal digits of a non-negative integral number in reverse order. More efficient than decimalDigits.

hexadecimalDigits :: Integral a => a -> Unfoldr a

Hexadecimal digits of a non-negative number.

reverseHexadecimalDigits :: Integral a => a -> Unfoldr a

Hexadecimal digits of a non-negative number in reverse order.

zipWithIndex :: Unfoldr a -> Unfoldr (Int, a)

Lift into an unfold, which produces pairs with index.

zipWithReverseIndex :: Unfoldr a -> Unfoldr (Int, a)

Deprecated: This function builds up stack. Use zipWithIndex instead.

Lift into an unfold, which produces pairs with right-associative index.

setBitIndices :: FiniteBits a => a -> Unfoldr Int

Indices of set bits.

unsetBitIndices :: FiniteBits a => a -> Unfoldr Int

Indices of unset bits.

textChars :: Text -> Unfoldr Char

Reproduces the behaviour of unpack.

Implementation is efficient and avoids allocation of an intermediate list.

textWords :: Text -> Unfoldr Text

Reproduces the behaviour of words.

Implementation is efficient and avoids allocation of an intermediate list.

trimWhitespace :: Unfoldr Char -> Unfoldr Char

Transformer of chars, replaces all space-like chars with space, all newline-like chars with \n, and trims their duplicate sequences to single-char. Oh yeah, it also trims whitespace from beginning and end.